+++ src/minter.js
@@ -7,7 +7,7 @@ import PostSignedTx from './api/post-signed-tx.js';
 import EstimateCoinSell from './api/estimate-coin-sell.js';
 import EstimateCoinBuy from './api/estimate-coin-buy.js';
 import EstimateTxCommission from './api/estimate-tx-commission.js';
-import ReplaceCoinSymbol, {ReplaceCoinSymbolByPath} from './api/replace-coin-symbol.js';
+import {ReplaceCoinSymbol, ReplaceCoinSymbolByPath} from './api/replace-coin.js';
 
 /**
  * @param {Object} [options]